requirements.txt
Flask==3.0.1
fpdf2==2.7.7
pillow==10.2.0
PyPDF2==3.0.1
pytesseract==0.3.10
qrcode==7.4.2

